Split up the properties in the test suite into submodules: fixes #272
* Split happens on the type of vector
* Discovered that only inlining of Unboxed vector functions is affected.
* Disable inlining on `testPolymorphicFunctions` for Unboxed to workaround
  the blowup.
